api/requirements.txt
Flask==2.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.1
Werkzeug==2.2.1
connexion[flask]==3.0.6
PyJWT==1.7.1
